refactor code actions
