+++ src/template-linter.ts
@@ -1,4 +1,5 @@
-import { Diagnostic, Files, TextDocument } from 'vscode-languageserver';
+import { Diagnostic, Files } from 'vscode-languageserver';
+import { TextDocument } from 'vscode-languageserver-textdocument';
 import { getExtension } from './utils/file-extension';
 import { toDiagnostic, toHbsSource } from './utils/diagnostic';
 import { getTemplateNodes } from '@lifeart/ember-extract-inline-templates';
